+++ src/taskfile-launcher/settings.ts
@@ -11,3 +11,15 @@ export function isResultExpanded(): boolean {
     .getConfiguration()
     .get('taskfileLauncher.defaultResultExpanded', false);
 }
+
+export function getShellPath(): string | undefined {
+  return vscode.workspace.getConfiguration().get('taskfileLauncher.shellPath');
+}
+
+export function getShellArgs(): string[] | undefined {
+  return vscode.workspace
+    .getConfiguration()
+    .get<string | undefined>('taskfileLauncher.shellArgs')
+    ?.split(' ')
+    .filter(Boolean);
+}